FAIT-Calculators/Home-Depot-Scraper/requirements_stripe_integration.txt
# Core dependencies
requests>=2.25.0
python-dotenv>=0.19.0
Flask==2.0.1
gunicorn==20.1.0
Jinja2==3.0.1
itsdangerous==2.0.1
werkzeug==2.0.1
click>=8.0.0
MarkupSafe>=2.0.0

# Stripe integration
stripe==5.4.0

# FastAPI services
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7

# Google Cloud
google-cloud-firestore==2.11.1

# Additional utilities
aiofiles==23.1.0
